demo: Hopsworks Tiktok RecSys pipeline
This pipeline might not be a 1 to 1 translation of the Flink version.
It should be possible to optimize the queries to be more performant.

Actual Flink Pipeline can be found here:
https://github.com/davitbzh/tiktok-recsys/blob/main/java/src/main/java/ai/hopsworks/tutorials/flink/tiktok/pipelines/TikTokStreamFeatureAggregations.java

The generator we use is taken from:
https://github.com/davitbzh/tiktok-recsys/tree/main/python/Jupyter/features

It was difficult to import it as a dependency, therefore copied it over.

Signed-off-by: Abhinav Gyawali <22275402+abhizer@users.noreply.github.com>